checkpoint - login works
You can signup, login, and logout. A cookie will be set when
authenticated.

Forms will get validated by the server.

There are examples of redirecting based off "authenticated" or not,
which currently just means the cookie is set. TODO is to think more
critically about where to store session state.